Update:
Fixed a problem when long_path was selected false: the module file was
incorrected copied as "module.in" into a directory called "default",
where the hidden version.in file was also copied. This is incorrect
because now "default" is the name of the module file, not of the
directory the module file is in. Also, probably we shouldn't create the
.version file in this case, which is a custom user installation.

It is probably also not a good idea to set PATH_ID to "default" in this
case. It is custom, not default.