Bugfix:
as for the Green's function we let the user pass a custom array of
frequencies for the susceptibilities. However, the old loop was still
behaving as if the first frequency was always 0+i0, evaluating it on its
onw and then starting the loop from the second one.

This is dangerous because

- the user could pass an array that starts from the n-th Matsubara freq
- the user could pass an array that is symmetric around 0 (TRIQS
  convention)

Uniformed the loop to the one for the other axes