Fix excitonic susceptibility for iorb=jorb
In the singlet and triplet_z component the condition was for ksector and
jsector to be both nonzero in order for the calculation to proceed.
This was too restrictive: sometimes one c+c operator can be applied but
the other not. In fact in triplet_xy ksector and jsector are evaluated
for each channel independently.

Written an auxiliary function `auxiliary_vvinit_manipulation` which
takes as input iorb, jorb, ksector, jsector, ichan, indx.

It defines isign and the sign of vup \pm vdw, tries to apply the
operators, does the tridiagonalization and add_to_lanczos and cleans up.